performance_testing/requirements.txt
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
matplotlib>=3.8.0
numpy>=1.24.0
